Fixed `getAdvice`
`getAdvice` trying to find unvisited node even if all nodes are visited and the process of searching takes too much time and memory in that case (because we have upper to 100! ways for 1 cell to visit another one). This cause because `findUnvisitedNode` try to find nearest neighbor node (but not from node list because it may be too far)